+++ tasks/ec2_deploy.js
@@ -53,14 +53,17 @@ module.exports = function(grunt){
             function deploy () {
                 var dest = util.format('%s/v/%s', root, v);
                 var target = root + '/current';
+                var stopped = '[[ $(pm2 jlist) == "[]" ]]';
+
                 var commands = [
                     util.format('sudo cp -r %s %s', remoteSync, dest),
                     util.format('sudo rm -rf `ls -t %s | tail -n +11`', root + '/v'),
                     util.format('sudo npm --prefix %s install --production', dest),
-                    util.format('sudo ln -sfn %s %s', dest, target), [
-                        util.format('sudo %s pm2 start %s/%s -i 2 --name %s', env, target, conf('NODE_SCRIPT'), name),
-                        'sudo pm2 reload all'
-                    ].join(' || ') // start or reload
+                    util.format('sudo ln -sfn %s %s', dest, target),
+                    util.format('%s || sudo pm2 reload all', stopped),
+                    util.format('%s && sudo %s pm2 start %s/%s -i 2 --name %s',
+                        stopped, env, target, conf('NODE_SCRIPT'), name
+                    )
                 ];
                 ssh(commands, name, log);
             }