+++ CHANGELOG.markdown
@@ -2,7 +2,8 @@
 
 - Sets `NODE_ENV` to the name of the EC2 instance
 - Added `ec2_reboot:name` task
-- 
+- Log HTTP url after deploys
+- Fixed `pm2` start/reload command
 
 # 0.2.7 Sun Bed